add TracesSearch highlight group
Create a new highlight group to apply to current search pattern. For
colorschemes with a subtle Search highlight group, it can be nice to use
a more distinct highlight for many of traces' operations. This is
especially true if the Search group is similar to the Visual group, as
matches are hard to see when searching within a range.